requirements.txt
aiogram==3.26.0
aiosqlite==0.22.1
python-dotenv==1.2.2
